Refactored autocomplete item generation
Now uses a better method of determining whether an item is sunset ("shelved" in BNet parlance). We examine the power cap for the item and use the watermark also contained within the quality block. Using iconWatermarkShelved is unreliable and can show e.g. exotics as sunset if they are from a season that has been sunset.